Basic functions defs moved to own file
